fix: move createFrom to concrete classes
