+++ packages/spacecat-shared-slack-client/src/index.d.ts
@@ -10,8 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-import type { UniversalContext } from '@adobe/helix-universal';
-
 import type { BaseSlackClient, ElevatedSlackClient } from './clients';
 import type { SlackChannel, SlackUser } from './models';
 
@@ -19,23 +17,6 @@ export declare const SLACK_TARGETS: {
   ADOBE_INTERNAL: string;
   ADOBE_EXTERNAL: string;
 };
-
-/**
- * Creates a slack client from the given context and for the given target.
- * A client can be designated as elevated, which allows it to perform advanced
- * operations such as user lookups by email and channel management.
- *
- * @param {UniversalContext} context The context to use to create the client.
- * @param {string} target The target to create the client for. See {@link SLACK_TARGETS}.
- * @param {boolean} [isElevated] Optional. Whether the client should be elevated. Default is false.
- */
-export declare function createFrom(
-  context: UniversalContext,
-  target: string,
-  isElevated?: boolean
-): BaseSlackClient | ElevatedSlackClient;
-
-export default createFrom;
 
 export {
   BaseSlackClient,
